test: use esm for prepare script
