+++ test/package.json
@@ -5,7 +5,7 @@
     "test": "jest",
     "g:ts-node": "cd $INIT_CWD && ts-node",
     "add-package": "yarn ../ pack && yarn add ../package.tgz",
-    "postinstall": "node prepare.js"
+    "postinstall": "node prepare.mjs"
   },
   "devDependencies": {
     "@nrwl/cli": "^15.9.7",
