+++ test/tsconfig.json
@@ -3,6 +3,7 @@
   "include": ["tests", "utils"],
 
   "compilerOptions": {
+    "types": ["jest", "node"],
     "noEmit": true,
     "strict": true,
     "esModuleInterop": true
